Add a basic `symex-define-command` macro
This is intended to encapsulate any setup and teardown actions that
in general would need to be performed with any symex transformation,
such as re-indenting, whitespace elimination, etc. It should also
avoid the need to rely on advice for some of these actions, and also
therefore make it easier for developers (including third parties) to
define such transformations, as all such logistical concerns can be
taken care of internally in the macro.