Fix an old bug in composing computations
Computations in practice are just identity transformations and the
results are lists that are composed by concatenation. But in some
cases the results are moves / traversals that need to be "lifted" to
be a list, and this was not happening, resulting in flatted traversal
specifications being appended at the end of some traversal results as
if they were lists. In particular, this was happening for circuits
with an unbounded number of repetitions, where the last repetition
resulted in (move 0 0) being flattened and appended to the result.

We ordinarily use the traversal result (as processed by the
computation) only to check success or failure, so this hasn't affected
the functioning of the traversals, but if wanted to, say, check the
length of the returned traversal, we might have gotten a higher number
than the real traversal length (e.g. (... move 0 0) would have counted
as three.