Improve traversal specifications to be more explicit
Traversal computations require a full account of traversals in order
to be accurate, and it turns out the traversal implementations thus
far were losing some information causing computations (which we hadn't
been using extensively up to this point) to be unreliable.

Specifically, `(move down)` is implemented as a primitive
motion. Unfortunately, this loses some information about the traversal
since we need to traverse some distance to get to the end of a branch,
but can get to its base in a single step, which obscures the fact that
we need to traverse back across the distance we covered to get across
the branch. This modifies the basic preorder/postorder traversals to
use a more well-formed "go down" traversal which retains a full
account, so they can be reliably used in traversal computations.